+++ lib/src/api/bitlabs_api.dart
@@ -10,30 +10,28 @@ class BitLabsApi {
   BitLabsApi(String token, String uid)
       : _headers = {'X-Api-Token': token, 'X-User-Id': uid};
 
-  Future<Response> checkSurveys() => get(url('check'), headers: _headers);
+  Future<Response> getSurveys() =>
+      get(url('v2/client/surveys', {'os': platform}), headers: _headers);
 
-  Future<Response> getActions() =>
-      get(url('actions', {'os': platform}), headers: _headers);
-
-  Future<Response> getOffers() => get(url('offers'), headers: _headers);
+  Future<Response> getOffers() =>
+      get(url('v2/client/offers'), headers: _headers);
 
   Future<Response> getLeaderboard() =>
-      get(url('leaderboard'), headers: _headers);
+      get(url('v1/client/leaderboard'), headers: _headers);
 
-  Future<Response> leaveSurveys(
-    String networkId,
-    String surveyId,
+  Future<Response> updateClick(
+    String clickId,
     String reason,
   ) {
-    return post(
-      url('networks/$networkId/surveys/$surveyId/leave'),
-      headers: {..._headers},
-      body: jsonEncode({'reason': reason}),
-    );
+    return post(url('v2/client/clicks/$clickId'),
+        headers: {..._headers},
+        body: jsonEncode({
+          'leave_survey': {'reason': reason}
+        }));
   }
 
   Future<Response> getAppSettings() =>
-      get(url('settings/v2'), headers: _headers);
+      get(url('v1/client/settings/v2'), headers: _headers);
 
   static Future<Response> getCurrencyIcon(String url) => get(Uri.parse(url));
 }